work to incorporate the nic of the hostapd events
I believe the nic can be used to determine which band
(2.4, 5, or 6) the client is connected to. It is probably different
per API but it's better than nothing.

Due to the abstractions I used previously this should be API-compatible.
Not that this matters as this system likely only works for my particular
setup.